+++ src/common/index.ts
@@ -1,4 +1,5 @@
 export * from "./configuration";
+export * from "./xqlint";
 export * from "./create-document-selector";
 export * from "./extension-state";
 export * from "./native-commands";