+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.1.2 (2023-02-26)
+* update to xqlint 0.2.1
+* add setting platform
+* common XQlint configuration.
+
 # 0.1.1 (2023-02-24)
 * Add setting to control "Hover show"
 * Add setting to filter out matching diagnostic messages